+++ lib/src/extensions/listenable_extension.dart
@@ -2,23 +2,20 @@
 
 import 'package:flutter/foundation.dart';
 
-/// Extension utility for [Listenable].
+///
 extension ListenableSyncExtension<T extends Listenable> on T {
-  /// Syncs notifications of this [Listenable] to a [notifier].
-  ///
-  /// This is a one-way sync. Changes to [notifier] will not be reflected in
-  /// this [Listenable].
-  ///
-  /// Useful for syncing any [Listenable] to a [ChangeNotifier].
-  ///
-  /// Example:
-  /// ```dart
-  /// class Counter extends ChangeNotifier {
-  ///   late final _count = ValueNotifier(0).sync(this);
-  /// }
-  /// ```
-  T sync(ChangeNotifier notifier) {
-    addListener(notifier.notifyListeners);
+  /// Sends notifications to a [ChangeNotifier].
+  T operator >>(ChangeNotifier other) {
+    addListener(other.notifyListeners);
+    return this;
+  }
+}
+
+///
+extension ChangeNotifierSyncExtension<T extends ChangeNotifier> on T {
+  /// Receives notifications from a [Listenable].
+  T operator <<(Listenable other) {
+    other.addListener(notifyListeners);
     return this;
   }
 }